# api/openapi-v2.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Windshift API v2",
    "version": "2.0.0",
    "description": "Canonical API v2. Browser clients use the session-authenticated /api/v2 mount; non-browser clients use this bearer-authenticated mount."
  },
  "servers": [
    {"url": "/rest/api/v2"}
  ],
  "paths": {
...
    "/time/worklogs": {
      "get": {
        "operationId": "listWorklogs",
        "security": [{"BearerAuth": ["time:read"]}],
        "parameters": [{"$ref": "#/components/parameters/Page"}, {"$ref": "#/components/parameters/PageSize"}],
        "responses": {"200": {"$ref": "#/components/responses/PageResponse"}, "400": {"$ref": "#/components/responses/InvalidRequestError"}}
      },
      "post": {
        "operationId": "createWorklog",
        "security": [{"BearerAuth": ["time:write"]}],
        "responses": {"201": {"$ref": "#/components/responses/DataResponse"}, "400": {"$ref": "#/components/responses/InvalidRequestError"}, "404": {"$ref": "#/components/responses/NotFoundError"}}
      }
    },
...
    "/time/worklogs/{worklog_id}": {
      "get": {
        "operationId": "getWorklog",
        "security": [{"BearerAuth": ["time:read"]}],
        "parameters": [{"$ref": "#/components/parameters/WorklogID"}],
        "responses": {"200": {"$ref": "#/components/responses/DataResponse"}, "404": {"$ref": "#/components/responses/NotFoundError"}}
      },
      "patch": {
        "operationId": "updateWorklog",
        "security": [{"BearerAuth": ["time:write"]}],
        "parameters": [{"$ref": "#/components/parameters/WorklogID"}],
        "responses": {"200": {"$ref": "#/components/responses/DataResponse"}, "400": {"$ref": "#/components/responses/InvalidRequestError"}, "404": {"$ref": "#/components/responses/NotFoundError"}}
      },
      "delete": {
        "operationId": "deleteWorklog",
        "security": [{"BearerAuth": ["time:delete"]}],
        "parameters": [{"$ref": "#/components/parameters/WorklogID"}],
        "responses": {"204": {"description": "Deleted"}, "404": {"$ref": "#/components/responses/NotFoundError"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "crw_*"
      }
    },
    "parameters": {
...
      ,"WorklogID": {"name": "worklog_id", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 1}}
...
      ,"Page": {"name": "page", "in": "query", "schema": {"type": "integer", "minimum": 1, "default": 1}}
      ,"PageSize": {"name": "page_size", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 50}}
    },
    "schemas": {
...
      "ErrorContent": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "details": {}
        }
      },
      "ErrorDocument": {
        "type": "object",
        "required": ["error", "request_id"],
        "properties": {
          "error": {"$ref": "#/components/schemas/ErrorContent"},
          "request_id": {"type": "string"}
        }
      },
      "DataDocument": {
        "type": "object",
        "required": ["data"],
        "properties": {"data": {}}
      },
      "PageDocument": {
        "type": "object",
        "required": ["data", "pagination"],
        "properties": {
          "data": {"type": "array", "items": {}},
          "pagination": {
            "type": "object",
            "required": ["page", "page_size", "total_items", "total_pages"],
            "properties": {
              "page": {"type": "integer"}, "page_size": {"type": "integer"},
              "total_items": {"type": "integer"}, "total_pages": {"type": "integer"}
            }
          }
        }
      }
...
    },
    "responses": {
      "DataResponse": {
        "description": "Canonical data document",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DataDocument"}}}
      },
      "PageResponse": {
        "description": "Canonical page document",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PageDocument"}}}
      },
...
      "NotFoundError": {
        "description": "Resource not found",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorDocument"}}}
      },
      "InvalidRequestError": {
        "description": "Request validation failed",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorDocument"}}}
      },
...
    }
  }
}